+++ docs/whats-new.md
@@ -29,6 +29,10 @@
 
 Release Date: TBD
 
+**`@math.gl/web-mercator`
+
+- `getBounds()` now supports an optional `fovy` parameter on the `WebMercatorViewport` object
+
 **`@math.gl/polygon`**
 
 - Improve performance of `getPolygonSignedArea()` by 3x